logging: Introduce NETWORKIT_RELEASE_LOGGING
Introduce a preprocessor definition that can be set to disable the
compilation of logging calls for levels TRACE and DEBUG.

Adjust CMakeLists.txt to automatically set this behavior in Release
mode and introduce a configuration option NETWORKIT_RELEASE_LOGGING
to control it. Using this option, users are still able to configure
release builds with full logging (e.g., to verify that a new
feature works correctly in release mode).